Backend/clubmg_bak_20190726/requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto3==1.9.133
botocore==1.12.134
celery==3.1.26.post2
Django==1.11.16
django-celery==3.2.2
django-filter==2.0.0
django-redis==4.10.0
djangorestframework==3.9.0
docutils==0.14
ffmpeg-python==0.1.17
functools32==3.2.3.post2
future==0.17.1
futures==3.2.0
jmespath==0.9.4
jsonschema==2.6.0
kombu==3.0.37
Markdown==3.0.1
MySQL-python==1.2.5
numpy==1.16.3
opencv-python==4.1.0.25
Pillow==5.3.0
PyJWT==1.6.4
python-dateutil==2.8.0
pytz==2018.7
redis==2.10.6
s3transfer==0.2.0
six==1.12.0
urllib3==1.24.2
uWSGI==2.0.17.1
